Successful extension build
