--- .github/workflows/submodule.ps1
@@ -1,2 +1,0 @@
-
-Invoke-Expression "git submodule update --init --recursive"
