+++ config.w32
@@ -1,0 +1,14 @@
+// $Id$
+// vim:ft=javascript
+
+ARG_ENABLE("noiselib", "Enable Noise extension features", "no");
+
+if (PHP_NOISELIB != "no") {
+    EXTENSION("noiselib", "noiselib.cpp", PHP_NOISELIB_SHARED, "/DZEND_ENABLE_STATIC_TSRMLS_CACHE=1 /I" + configure_module_dirname);
+    ADD_FLAG("CFLAGS_NOISELIB", "/EHsc");
+    ADD_SOURCES(
+        configure_module_dirname + "/src",
+        "Noise.cpp",
+        "noiselib"
+    );
+}
