--- noisegen.ini
@@ -1,1 +1,0 @@
-extension=noisegen.so
